refactor: componentsディレクトリを新設し、コンポーネントファイルを整理。理由: ファイル構成の明確化・保守性向上のため。
