+++ src/components/StackedImage.js
